Fix inconsistent state of wizard forms
Dispatching `initNewMovement` after mount led to problems, because
`destroy('wizard')` is dispatched in the saga which is triggered
(i.e. `initNewDeparture`).
The aim of the `destroy` action in the initialization saga is to start with
a fresh wizard form, no matter what happened before. However, if we do it in
`componentDidMount`, it is executed after `redux-form/INITIALIZE`,
which is certainly not what we want to achieve. If we do it in
`componentWillMount` instead, `redux-form/DESTROY` is executed right before
`redux-form/INITIALIZE` and everything works fine.