Temporarily deactivate validation for arrivalRoute and landingCount
There is a new bug since 5fc36fc:

If wizard is opened for existing departure or arrival or a new departure is
created from an existing arrival or the other way round, the click on the "next"
button to proceed to the pilot page is ignored. One has to focus the aircraft
registration field to "activate" the "next" button.

I'm not able to reproduce this bug on my local environment, so I have to deploy
this new version to check it on mfgt-flights-redux.firebaseapp.com.