Don't render wizard form before it is fully intitialized
This should fix the bug which is mentioned in commit `f3eae68`.

Before this change, we immediately started with rendering the form in
`<MovementWizard/>`. During rendering, redux form dispatched `REGISTER_FIELD`
actions which initializes the form for validation and stuff like this. During
the rendering we dispatched `destroy('wizard')'` in our shared movement sagas
to start with a clean form. But in fact it messed up the whole form state,
because we already started with rendering and therefore with initialization
too early.

Now, we wait until the `initialized` flag of the wizard state is set to true.
This is set to true in the shared movement sagas after `destroy`
and `initialize`.