Only validate values which are visible in the current wizard step
This fixes the bug which is mentioned in commit f3eae68.

We should only validate those values which are visible in the current wizard
page. Otherwise redux-form validates the initial values on form initialization
before any fields are registered (which will be registered when they're getting
rendered). As a result the validation fails for some values which are not even
visible in the first wizard step and it's not possible to move on to the next
page because of the negative validation result (so we're stuck on the first
page and can't do anything there).