Fix layout issues in movement wizard
Since commit 032cef2 ("Fix layout and infinite scroll issues in
VerticalHeaderLayout"), the wizard was too wide (the wizard form took the full
width of the window and there was also a vertical header on the left side which
needed some space (17%). Now, the wizard form takes the window width minus the
width of the vertical header - just like before commit 032cef2.