Keep Firebase listeners when Firebase session expires
Until now, we added `child_added`, `child_changed` and `child_removed`
listeners on the Firebase query in remote#loadLimited. These listeners were
supposed to live forever to receive all subsequent events on the loaded range.

Albeit supposed to live forever, Firebase cancelled the listeners once the
current session expired (which is the case after 24h). The listeners get
removed, because the client is not allowed to access the node for a short
amount of time, until the session has been refreshed with a new token.

Fix:

To ensure that there are listeners for the loaded ranges, we keep the query
references in our state and register the listeners in
MovementList#componentWillMount() (which will be called after the session has
been refreshed).
Also, the listeners are still getting registered directly after a range has been
loaded using `remote.loadLimited` (see `loadMovements` saga).