Fix scroll-to-top bug after react-router update
There was an `onChange` handler on the root route (which rendered the
<App> component) before the react-router update to scroll to the top.
when the location changes.
The event handler hasn't been migrated to the new API so far.
Since react-router routes are real React components now, the React
lifecycle methods can be used now - `componentWillReceiveProps` in
this case (there is no onChange event and the like anymore on the
react-router components (i.e. Route)).