Add more end to end tests (for start page)
We have to disable IP authentication in the end to end tests, because
for some reason the `auth` fetch gets cancelled when running the
tests with `npm run cy:run`. Once a fetch got cancelled, the tests
don't continue. Couldn't figure out why this is happening - so I de-
activated the IP authentication in the tests.