Add possibility to disable links in UI for certain users
If the Flightbox runs at the airport on a terminal with a special
static login it's not desired that anyone can log out or leave
the browser window.

For these logins, links can be disabled in the database by setting
`logins[$login_id].links` to `false`.

This hides the "logout" link at the top right and also disables the
"open digital Flightbox" link at the bottom right.

Note that the admin collection has also been moved to the `logins`
collection. So, `admins[$login_id]: true` has to be changed to
`logins[$login_id].admin: true` in the database.

Ticket: https://app.asana.com/0/1201466020311786/1201613928400936