Take custom aircrafts from settings into account during import
You can specify some custom aircrafts under /settings/aircrafts/custom
that should be part of the aircrafts list even if they're not part of
the CSV that is imported. E.g. that's useful if there are some aircrafts
home based on your airport that are registered abroad (e.g. aircrafts
with "N...." registration on an airport in Switzerland, where we usually
import only the aircrafts from the BAZL register ("HB...").