Remove annoying automatic scrolling in dropdown component
The automatic scrolling is only desired when navigating with the arrow
keys on the keyboard. In this case, we keep the scrolling. However,
it's not desired on every component update (which also occurs when
we move the cursor over the list of options -> no automatic scolling
desired in this case).