Fix movement association if filtered with circuits
Example movement sequence of *non-homebased* aircraft:
1. arrival
2. circuits departure
3. circuits arrival
4. departure

Here, movements 1 and 4 are associated and movements 2 and 3.
However if we set the date filtering so that movements 1 to 3
were visible or 2 to 4, the matching counterpart for movement 1
resp. 4 couldn't be found. The reason was that we didn't filter
out the circuit movements in between in this case. Now the issue
is fixed.