Filter out corrupt movements in movement list
Sometimes there's a bug which results in empty departures/arrivals with
just `associatedMovement: {type: 'none' }`.

Couldn't find out what the reason for this yet. But by filtering them
out here we make sure that the movement list can still be used (otherwise
there are null pointers, because all the required fields are missing on
the movement).