Add `local` flag to auth data if user has logged in on site
This is the case, if the user came to the Flightbox by scanning
a QR code which opens the Flightbox URL with a query token which
also allows to sign in as guest.

The `local` flag is true if the `uid` is `guest` (which can only be the
case if they came from a QR code link) or if the email login was
executed with a query token being present.