Integrate Sentry.io to capture client errors
There is a LSZM user who isn't able to sign using his iPad mini.
I'm haven't been able to reproduce the problem yet. Maybe there is
a client error that's being logged now that's gonna be helpful to
fix the issue.

Note that we have to use the version 6.2.4 because the newer ones aren't
compatible with the versions we use in our tech stack currently.