Show "Zahlung offen" tag and payment link if no payment method set
After saving an arrival, the payment screen is shown to the user. Some
users seem to leave this screen without paying (or setting the payment
method "invoice"). We now mark those arrivals accordingly where this
had been the case. There is also a link shown on the arrival details
for these arrivals to return to the payment screen.

Note that if a payment method like "checkout" or "card" is selected,
they could still have cancelled the payment and the payment could
still be open, even though we don't show "Zahlung offen" in this case,
because the payment method would be set. This is something we will
address later.